refactors lib folder and updates imports accordingly
