+++ src/lib/ui/icons/devops.svelte
