+++ src/lib/ui/icons/email.svelte
