+++ src/lib/ui/icons/external.svelte
