+++ src/lib/ui/icons/go.svelte
