+++ src/lib/ui/icons/graphql.svelte
