+++ src/lib/ui/icons/node.svelte
