+++ src/lib/ui/icons/rust.svelte
