+++ src/lib/ui/icons/svg-colors.scss
@@ -1,4 +1,4 @@
-@import '../../style/variables.scss';
+@import '../../../style/variables.scss';
 svg {
 	&.none {
 		fill: none;
