+++ src/lib/ui/loading/SkeletonImage.svelte
